Initial import of the (non-working) app.
The application shows two squares that are connected to the edges
of the diagram by lines. When the user drags a square with the mouse,
it should move and the lines should update too.

But this is currently not working, because the app violates a crucial
rule of redux: The reducer should return a new state object, but right
now, it does not.