Added immutable records for nodes and edges.
Immutable records simplify access to the properties of the objects.
They *also* make sure that one cannot set any properties that should
not be inside the record.

The reducer works exactly as before, only the initial state adds
records to the nodes and edges arrays (instead of maps). The Link
component is somewhat simpler now: It can access the properties of
the records directly, instead of using "get(...)".